wip: NamespacePrefixMapper in a dedicated file
